docs: various fixes
