+++ packages/pure-parse/docs/guide/comparison.md
@@ -1,0 +1,5 @@
+# Comparison With Other Libraries
+
+`pure-parse` is benchmarked against other validation libraries:
+
+- It's roughtly 4× faster than Zod
