+++ packages/pure-parse/src/common/index.ts
@@ -1,2 +1,3 @@
 export * from './Primitive'
 export * from './Infer'
+export * from './json'
