+++ packages/pure-parse/src/common/json.ts
@@ -1,0 +1,10 @@
+/**
+ * A value that represent any JSON-serializable data
+ */
+export type JsonValue =
+  | null
+  | boolean
+  | number
+  | string
+  | { [x: string]: JsonValue }
+  | JsonValue[]
