+++ packages/pure-parse/src/guard/index.ts
@@ -1,3 +1,2 @@
 export * from './validation'
-export * from './json'
 export * from './guards'
