+++ packages/pure-parse/src/guard/json.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from 'vitest'
-import { isJsonValue, parseJson } from './json'
-import { object } from './validation'
-import { isNumber, isString, isUnknown } from './guards'
+import { isJsonValue } from './json'
+import { object, isNumber, isString, isUnknown } from './index'
 
 describe('json', () => {
   describe('validation', () => {
@@ -43,24 +42,6 @@ describe('json', () => {
     })
     it('is recursive arrays and objects', () => {
       expect(isJsonValue(JSON.stringify([{ a: [{ b: [] }] }]))).toEqual(true)
-    })
-  })
-  describe('parseJson', () => {
-    it('does not throw errors', () => {
-      expect(() => parseJson(isUnknown)('not a json!')).not.toThrow()
-    })
-    it('returns an error if the parsing failed', () => {
-      expect(parseJson(isUnknown)('not a json!')).toBeInstanceOf(Error)
-    })
-    it('returns an error if the validation failed', () => {
-      expect(parseJson(isString)('{}')).toBeInstanceOf(Error)
-    })
-    it('returns the object if the validation succeeded', () => {
-      expect(parseJson(isString)(JSON.stringify('abc'))).toEqual('abc')
-      expect(parseJson(isNumber)(JSON.stringify(123))).toEqual(123)
-      expect(
-        parseJson(object({ a: isNumber }))(JSON.stringify({ a: 1 })),
-      ).toEqual({ a: 1 })
     })
   })
 })
