+++ packages/pure-parse/src/guard/json.ts
@@ -1,16 +1,6 @@
-import { array, partialRecord, union, Guard } from './validation'
+import { array, partialRecord, union } from './validation'
 import { isBoolean, isNull, isNumber, isString } from './guards'
-
-/**
- * A value that represent any JSON-serializable data
- */
-export type JsonValue =
-  | null
-  | boolean
-  | number
-  | string
-  | { [x: string]: JsonValue }
-  | JsonValue[]
+import { JsonValue } from '../common'
 
 export const isJsonValue = (data: unknown): data is JsonValue =>
   union(
@@ -21,18 +11,3 @@ export const isJsonValue = (data: unknown): data is JsonValue =>
     partialRecord(isString, isJsonValue),
     array(isJsonValue),
   )(data)
-
-/**
- * This function will not throw an error.
- * @param is - a validation function
- */
-export const parseJson =
-  <T>(is: Guard<T>) =>
-  (text: string): T | Error => {
-    try {
-      const data = JSON.parse(text)
-      return is(data) ? data : new Error('Validation failed')
-    } catch (e) {
-      return e instanceof Error ? e : new Error(`Unknown error: ${e}`)
-    }
-  }
